lib/utils Ordner aufgeräumt (inkl. imports)
