+++ lib/screens/settings/manage_profiles/profile_list_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:glutter/screens/settings/manage_profiles/profile_create_screen.dart';
 import 'package:glutter/screens/settings/manage_profiles/profile_edit_screen.dart';
 import 'package:glutter/services/shared/database_service.dart';
-import 'package:glutter/utils/toast.dart';
+import 'package:glutter/utils/utils.dart';
 
 class ProfileListScreen extends StatefulWidget {
     ProfileListScreen({Key key, this.title: "Profiles"}) : super(key: key);
