--- lib/utils/check_values.dart
@@ -1,12 +1,0 @@
-
-bool valuesHaveChanged(Map values) {
-
-    bool changed = false;
-    values.forEach((key, value) {
-        if(key != value) {
-            changed = true;
-        }
-    });
-
-    return changed;
-}
