--- lib/utils/launch_url.dart
@@ -1,9 +1,0 @@
-import 'package:url_launcher/url_launcher.dart';
-
-launchURL(url) async {
-    if (await canLaunch(url)) {
-        await launch(url);
-    } else {
-        throw 'Could not launch $url';
-    }
-}
